requirements.txt
slack_sdk==3.21.0
openai==1.3.0
flask==2.3.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
